requirements.txt
# requirements.txt
python-git
obsws-python
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
pywin32
